requirements.txt
APScheduler==3.7.0
asgiref==3.3.2
cffi==1.14.5
Django==3.2
django-redis==4.12.1
gevent==21.1.2
greenlet==1.0.0
pycparser==2.20
PyJWT==2.1.0
pyserial==3.5
pytz==2021.1
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
zope.event==4.5.0
zope.interface==5.3.0
